refactor: 사용자 이름 Name 클래스로 분리
